refactor: make `FinalRoundBuilder::make_sumcheck_polynomial` into standalone `make_sumcheck_polynomial`
This includes refactoring and moving a test
